LPS-182364 Regen

+++ modules/apps/blogs/blogs-uad/src/main/java/com/liferay/blogs/uad/anonymizer/BaseBlogsEntryUADAnonymizer.java
@@ -62,8 +62,12 @@ public abstract class BaseBlogsEntryUADAnonymizer
 	}
 
 	@Override
-	public void delete(BlogsEntry blogsEntry) throws PortalException {
-		blogsEntryLocalService.deleteBlogsEntry(blogsEntry);
+	public void delete(BlogsEntry blogsEntry, long userId)
+		throws PortalException {
+
+		if (blogsEntry.getUserId() == userId) {
+			blogsEntryLocalService.deleteBlogsEntry(blogsEntry);
+		}
 	}
 
 	@Override
