+++ modules/apps/bookmarks/bookmarks-uad/src/main/java/com/liferay/bookmarks/uad/anonymizer/BaseBookmarksFolderUADAnonymizer.java
@@ -62,8 +62,12 @@ public abstract class BaseBookmarksFolderUADAnonymizer
 	}
 
 	@Override
-	public void delete(BookmarksFolder bookmarksFolder) throws PortalException {
-		bookmarksFolderLocalService.deleteFolder(bookmarksFolder);
+	public void delete(BookmarksFolder bookmarksFolder, long userId)
+		throws PortalException {
+
+		if (bookmarksFolder.getUserId() == userId) {
+			bookmarksFolderLocalService.deleteFolder(bookmarksFolder);
+		}
 	}
 
 	@Override
