+++ modules/apps/calendar/calendar-uad/src/main/java/com/liferay/calendar/uad/anonymizer/BaseCalendarBookingUADAnonymizer.java
@@ -62,8 +62,12 @@ public abstract class BaseCalendarBookingUADAnonymizer
 	}
 
 	@Override
-	public void delete(CalendarBooking calendarBooking) throws PortalException {
-		calendarBookingLocalService.deleteCalendarBooking(calendarBooking);
+	public void delete(CalendarBooking calendarBooking, long userId)
+		throws PortalException {
+
+		if (calendarBooking.getUserId() == userId) {
+			calendarBookingLocalService.deleteCalendarBooking(calendarBooking);
+		}
 	}
 
 	@Override
