+++ modules/apps/dynamic-data-lists/dynamic-data-lists-uad/src/main/java/com/liferay/dynamic/data/lists/uad/anonymizer/BaseDDLRecordUADAnonymizer.java
@@ -62,8 +62,12 @@ public abstract class BaseDDLRecordUADAnonymizer
 	}
 
 	@Override
-	public void delete(DDLRecord ddlRecord) throws PortalException {
-		ddlRecordLocalService.deleteRecord(ddlRecord);
+	public void delete(DDLRecord ddlRecord, long userId)
+		throws PortalException {
+
+		if (ddlRecord.getUserId() == userId) {
+			ddlRecordLocalService.deleteRecord(ddlRecord);
+		}
 	}
 
 	@Override
