+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-uad/src/main/java/com/liferay/dynamic/data/mapping/uad/anonymizer/BaseDDMFormInstanceRecordUADAnonymizer.java
@@ -65,11 +65,13 @@ public abstract class BaseDDMFormInstanceRecordUADAnonymizer
 	}
 
 	@Override
-	public void delete(DDMFormInstanceRecord ddmFormInstanceRecord)
+	public void delete(DDMFormInstanceRecord ddmFormInstanceRecord, long userId)
 		throws PortalException {
 
-		ddmFormInstanceRecordLocalService.deleteFormInstanceRecord(
-			ddmFormInstanceRecord);
+		if (ddmFormInstanceRecord.getUserId() == userId) {
+			ddmFormInstanceRecordLocalService.deleteFormInstanceRecord(
+				ddmFormInstanceRecord);
+		}
 	}
 
 	@Override
