+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-uad/src/main/java/com/liferay/dynamic/data/mapping/uad/anonymizer/BaseDDMFormInstanceUADAnonymizer.java
@@ -62,8 +62,12 @@ public abstract class BaseDDMFormInstanceUADAnonymizer
 	}
 
 	@Override
-	public void delete(DDMFormInstance ddmFormInstance) throws PortalException {
-		ddmFormInstanceLocalService.deleteFormInstance(ddmFormInstance);
+	public void delete(DDMFormInstance ddmFormInstance, long userId)
+		throws PortalException {
+
+		if (ddmFormInstance.getUserId() == userId) {
+			ddmFormInstanceLocalService.deleteFormInstance(ddmFormInstance);
+		}
 	}
 
 	@Override
