+++ modules/apps/message-boards/message-boards-uad/src/main/java/com/liferay/message/boards/uad/anonymizer/BaseMBCategoryUADAnonymizer.java
@@ -62,8 +62,12 @@ public abstract class BaseMBCategoryUADAnonymizer
 	}
 
 	@Override
-	public void delete(MBCategory mbCategory) throws PortalException {
-		mbCategoryLocalService.deleteCategory(mbCategory);
+	public void delete(MBCategory mbCategory, long userId)
+		throws PortalException {
+
+		if (mbCategory.getUserId() == userId) {
+			mbCategoryLocalService.deleteCategory(mbCategory);
+		}
 	}
 
 	@Override
