+++ modules/apps/message-boards/message-boards-uad/src/main/java/com/liferay/message/boards/uad/anonymizer/BaseMBSuspiciousActivityUADAnonymizer.java
@@ -59,11 +59,13 @@ public abstract class BaseMBSuspiciousActivityUADAnonymizer
 	}
 
 	@Override
-	public void delete(MBSuspiciousActivity mbSuspiciousActivity)
+	public void delete(MBSuspiciousActivity mbSuspiciousActivity, long userId)
 		throws PortalException {
 
-		mbSuspiciousActivityLocalService.deleteMBSuspiciousActivity(
-			mbSuspiciousActivity);
+		if (mbSuspiciousActivity.getUserId() == userId) {
+			mbSuspiciousActivityLocalService.deleteMBSuspiciousActivity(
+				mbSuspiciousActivity);
+		}
 	}
 
 	@Override
