+++ modules/apps/portal-workflow/portal-workflow-kaleo-uad/src/main/java/com/liferay/portal/workflow/kaleo/uad/anonymizer/BaseKaleoTaskAssignmentInstanceUADAnonymizer.java
@@ -60,11 +60,15 @@ public abstract class BaseKaleoTaskAssignmentInstanceUADAnonymizer
 	}
 
 	@Override
-	public void delete(KaleoTaskAssignmentInstance kaleoTaskAssignmentInstance)
+	public void delete(
+			KaleoTaskAssignmentInstance kaleoTaskAssignmentInstance,
+			long userId)
 		throws PortalException {
 
-		kaleoTaskAssignmentInstanceLocalService.
-			deleteKaleoTaskAssignmentInstance(kaleoTaskAssignmentInstance);
+		if (kaleoTaskAssignmentInstance.getUserId() == userId) {
+			kaleoTaskAssignmentInstanceLocalService.
+				deleteKaleoTaskAssignmentInstance(kaleoTaskAssignmentInstance);
+		}
 	}
 
 	@Override
