+++ modules/apps/wiki/wiki-uad/src/main/java/com/liferay/wiki/uad/anonymizer/BaseWikiPageUADAnonymizer.java
@@ -62,8 +62,10 @@ public abstract class BaseWikiPageUADAnonymizer
 	}
 
 	@Override
-	public void delete(WikiPage wikiPage) throws PortalException {
-		wikiPageLocalService.deletePage(wikiPage);
+	public void delete(WikiPage wikiPage, long userId) throws PortalException {
+		if (wikiPage.getUserId() == userId) {
+			wikiPageLocalService.deletePage(wikiPage);
+		}
 	}
 
 	@Override
