+++ modules/util/portal-tools-service-builder-test-uad/src/main/java/com/liferay/portal/tools/service/builder/test/uad/anonymizer/BaseUADPartialEntryUADAnonymizer.java
@@ -57,8 +57,12 @@ public abstract class BaseUADPartialEntryUADAnonymizer
 	}
 
 	@Override
-	public void delete(UADPartialEntry uadPartialEntry) throws PortalException {
-		uadPartialEntryLocalService.deleteUADPartialEntry(uadPartialEntry);
+	public void delete(UADPartialEntry uadPartialEntry, long userId)
+		throws PortalException {
+
+		if (uadPartialEntry.getUserId() == userId) {
+			uadPartialEntryLocalService.deleteUADPartialEntry(uadPartialEntry);
+		}
 	}
 
 	@Override
